+++ c02_actix-gcd/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2024"
 [dependencies]
 actix-web = "4.10.2"
 serde = { version = "1.0.219", features = ["derive"] }
+c02_hello = { workspace = true }
