+++ c02_actix-gcd/src/main.rs
@@ -1,5 +1,6 @@
 use actix_web::{App, HttpResponse, HttpServer, Responder, get, post, web};
 use serde::Deserialize;
+use c02_hello::gcd;
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -49,17 +50,4 @@ async fn post_gcd(form: web::Form<GcdParameters>) -> impl Responder {
     );
 
     HttpResponse::Ok().content_type("text/html").body(response)
-}
-
-fn gcd(mut n: u64, mut m: u64) -> u64 {
-    assert!(n != 0 && m != 0);
-    while m != 0 {
-        if m < n {
-            let t = m;
-            m = n;
-            n = t;
-        }
-        m = m % n;
-    }
-    n
 }
