+++ c02_hello/src/main.rs
@@ -1,5 +1,6 @@
 use std::env;
 use std::str::FromStr;
+use c02_hello::gcd;
 
 fn main() {
     // println!("Hello, world!");
@@ -20,23 +21,4 @@ fn main() {
         d = gcd(d, *m);
     }
     println!("The greatest common divisor of {:?} is {}", numbers, d);
-}
-
-fn gcd(mut n: u64, mut m: u64) -> u64 {
-    assert!(n != 0 && m != 0);
-    while m != 0 {
-        if m < n {
-            let t = m;
-            m = n;
-            n = t;
-        }
-        m = m % n;
-    }
-    n
-}
-
-#[test]
-fn test_gcd() {
-    assert_eq!(gcd(14, 15), 1);
-    assert_eq!(gcd(2 * 3 * 5 * 11 * 17, 3 * 11 * 19), 3 * 11);
 }
